login: Ignore VPD error for device owners
As part of chromium:618340 the value zero has to be stored in VPD
for device owners. In case flashrom is corrupted, the update of
VPD fails, making user login impossible. The proposed fix,
implemented in this CL is to ignore the VPD update error in case
the device is not enrolled.

BUG=chromium:669520
TEST=Manually tested on a reks device.

Change-Id: I9c653b051fd5cb11e36d858a0d7e09c556f339f7
Reviewed-on: https://chromium-review.googlesource.com/414890
Commit-Ready: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Thiemo Nagel <tnagel@chromium.org>