libcontainer: fork/exec setfiles only once for faster ARC boot
Previously, run_setfiles_command() which spawns /sbin/setfiles was
called for each mount_external mount point and because of that,
labeling files and directories took 800-900ms on Cyan.

With this CL, /sbin/setfiles is called only once, and the time will
be like ~150ms.

BUG=b:33348324
TEST=ARC still starts

Change-Id: I2fa97cfb1d5920ef7fadbb3f289530baeb8bb2e4
Reviewed-on: https://chromium-review.googlesource.com/416394
Commit-Ready: Yusuke Sato <yusukes@chromium.org>
Tested-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>