power: Log a few more cases when the idle action is ignored.
Make powerd log an informative message when the default idle
action is ignored due to the disable_idle_suspend pref being
set (as happens automatically in dev/test images) or an
update being applied while on AC power.

These overrides happen when powerd is computing its
settings, so while it was possible to see that the
Chrome-supplied action was being ignored in the current
logs, the reason why was non-obvious.

BUG=chromium:666430
TEST=manual: used set_power_policy to set a 1m idle delay,
     then watched the logs and saw "Not performing idle
     action because disable_idle_suspend powerd pref is set
     (done automatically in dev)"

Change-Id: If95b78d2a74ed3645c6b10d0ea737b0af2653401
Reviewed-on: https://chromium-review.googlesource.com/416374
Commit-Ready: Dan Erat <derat@chromium.org>
Tested-by: Dan Erat <derat@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>