arc-networkd: Don't crash if rt_tables isn't populated right away
If the user logs in with no active network connection, arc-networkd
will check every few seconds to see if: 1) arc0 inside the container is
up, and 2) rt_tables has an entry for arc0.  (1) should happen in the
absence of a LAN connection, but (2) will not, causing arc-networkd
to time out and abort.

Move the check for (2) into the Set() function, which runs a couple
of seconds after the external LAN interface is up.

BUG=chromium:649247
TEST=`tail -f /var/log/net.log` with no connection, wifi connection

Change-Id: Ie2680c68810af9ff082a589e136ae98fbf7bd00c
Reviewed-on: https://chromium-review.googlesource.com/416362
Commit-Ready: Kevin Cernekee <cernekee@chromium.org>
Tested-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>