platform2: add .clang-format symlink
According to go/brillo-c++-style, the C++ style specified in
common-mk/brillo-clang-format applies to platform2 code. To help
sub-projects adopt the style and maintain consistency across the
platform2 code base, this CL adds .clang-format to the platform2 root
directory, so that all sub-projects can pick it up by default. Any
sub-project can still choose to provide a .clang-format override in a
sub-directory under platform2. The added .clang-format is currently a
symlinked to common-mk/brillo-clang-format. If we later decide to adopt
a different style, we can simply update the symlink to point to a
different clang-format config file, or turn it into a real clang-format
config file.

BUG=None
TEST=Run clang-format on code of a few sub-projects, with and without a local .clang-format file.
TEST=Run `clang-format --dump-config` under a platform2 sub-directory and verify the configuration.

Change-Id: I9330b2b69674af222c0c074a7777385d7486289c
Reviewed-on: https://chromium-review.googlesource.com/416194
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>