authpolicy: Use protobuf configuration file
Uses a dedicated configuration file in protobuf format instead of
parsing the smb.conf file to retrieve configuration. Writes smb.conf
to a temp directory every time it is used. This is more secure (protobuf
parsing is battle-tested) and it allows us to update the smb.conf, which
was impossible before since it was written once at domain join.

BUG=chromium:666693
TEST=Compiles, tested with custom test code.

Change-Id: I268e916293d49ff0faf31435a0e39c991790f383
Reviewed-on: https://chromium-review.googlesource.com/418995
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>