login: Split out wallpaper code and add test
We want to enhance the wallpaper flags code to check the model name at
run-time and use the correct wallpaper for that model. In preparation for
this, move this code into a separate function. Also add a few simple tests
to cover the current behaviour.

This should not change any logic.

BUG=chromium:671391
TEST=FEATURES=test emerge-reef chromeos-ss[plpoklogin
Signed-off-by: Simon Glass <sjg@chromium.org>

Change-Id: Icce2ac34f9b4f24b8763153048adc84d0b1b0a12
Reviewed-on: https://chromium-review.googlesource.com/419733
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>