userspace_touchpad: Temporary fake touchpad for device bringup
This is an userspace driver with very basic support for an experimental
touchpad.  The purpose is to read in touch events over i2c and convert
them into kernel events using the uinput library to create an virtual
input device.

This will be used for device bringup but will be replaced shortly.  In
the long run, we will be using a standardized driver like hid over i2c.
This is just a stop-gap solution to allow some touchpad functionality
during the very early stages of  bringup.

Big thanks to charliemooney@ who wrote everything outside of main.cc

BUG=chrome-os-partner:61088
TEST=Tested on an Eve with an attached st touchpad, evtest showed the
device and I was able to move the pointer around on the screen.
CQ-DEPEND=CL:423255

Change-Id: Idebeac4df96fcf96479f49de4f631e2200ce4fe4
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/420808
Commit-Ready: Tai-hsu Lin <sheckylin@google.com>
Tested-by: Tai-hsu Lin <sheckylin@google.com>
Tested-by: Wei-Ning Huang <wnhuang@chromium.org>
Reviewed-by: Tai-hsu Lin <sheckylin@google.com>
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>