shill: minor cleanup in dhcp_provider
This round-trip conversion through base::Process is useless because
Process doesn't do any checking to make sure the process is real
anyway, and the lifetime of the object doesn't have anything to do
with the lifetime of the process it represents anyway. Simplify
this away.

BUG=None
TEST=unit tests

Change-Id: I12c3e6cace258cbdf41f9732340934c80f1d7719
Reviewed-on: https://chromium-review.googlesource.com/422556
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>