init: Check for block_devmode=1 included when sysfs present
When checking the value of block_devmode for devices that allow
reading of VPD through sysfs, the old version was looking only at
the presence of the flag. This CL checks also the value of
block_devmode and only if it's 1, the devmode is blocked.

BUG=chromium:618340
TEST=Tested manually in console

Change-Id: I6c0eb33e0734b3d613959b6de9772d8cf2e40d55
Reviewed-on: https://chromium-review.googlesource.com/423048
Commit-Ready: Igor <igorcov@chromium.org>
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Thiemo Nagel <tnagel@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>