common-mk: import gyp compiler logic
Since the actual compiler code we need from gyp is small, import a
copy of it here so people don't have to install gyp itself.  This
shouldn't be a big deal since upstream gyp is a dead project now,
so it's unlikely that this code will ever change.

BUG=None
TEST=gyplinting still works

Change-Id: I2fcac0463f096127e89a7f81281656a7247a68ae
Reviewed-on: https://chromium-review.googlesource.com/423367
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>