authpolicy: Improve error messaging for expired passwords
When the password is expired and kinit fails after asking for the
old and twice the new password, authpolicy reports "password expired"
as error, even though the error might be different (e.g. a seccomp
failure as in my test case). This CL fixes this.

BUG=chromium:676648
TEST=Tested the case that failed before and it works fine now

Change-Id: Ie19c853cf7010e0fbdfc28dfcd5f3db34194bda9
Reviewed-on: https://chromium-review.googlesource.com/424396
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>