hal_adapter: restore buffer handle in process_capture_result
When processing a capture result, the frameworks expect to see the same
buffer handle (i.e. same pointer address) as the one passed down in
process_capture_request. When the capture request is sent from ARC to
CrOS, we need to pass the whole buffer handle to HAL adapter so the HAL
can get a fd to access the underlying buffer. On the other hand, when
CrOS sends the capture result to ARC, the Android HAL needs to restore
the original buffer_handle_t in the capture result before passing up to
the frameworks.

BUG=b:30824236
TEST=Manually tested with the arc-camera3-hal-test package.

Change-Id: Ibf64245a7007d301e70c94f13a00aca3fcb5e253
Reviewed-on: https://chromium-review.googlesource.com/424656
Commit-Ready: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>