biod: fix use after move of fpc image data
This bug slipped through because using a moved from vector is still
valid, but its contents are undefined. Old implementations of
bio_image_set_data did not check or use the data parameter, and so did
not complain when we gave it invalid data.

TEST=bio_client_tool enroll FpcBiometric
BUG=None

Change-Id: Icec6d271df216e51a859ad9e0a384f4ad27c747a
Reviewed-on: https://chromium-review.googlesource.com/425437
Commit-Ready: Mengqi Guo <mqg@chromium.org>
Tested-by: Mengqi Guo <mqg@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>