libcontainer: add schedtune to the list of used controllers
We need schedtune to be available in Android container (if it is
available on the host), so let's add it to the list of cgroup
controllers we create.

Also we need to recognize that not all cgroup controllers are available
on all platforms, so gracefully handle missing ones.

BUG=chrome-os-partner:59485
TEST=Build and boot on Kevin

Change-Id: I633d601b14163a4c25271c08afb8d53bbfbbee14
Reviewed-on: https://chromium-review.googlesource.com/425199
Commit-Ready: Dmitry Torokhov <dtor@chromium.org>
Tested-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>