metrics: Report basic crouton usage
Detects if /run/crouton exists, and sends "Platform.Crouton.Started"=1
when that happens.

Note that we also send "Platform.Crouton.Started"=0 on boot, for
accurate counting, so that the number of elements in "1" bucket
tells us about active users, and "0" tells us the total number
of users.

BUG=chromium:678865
TEST=emerge-elm metrics; cros deploy IP metrics; reboot
     check chrome://histograms after boot

Change-Id: I06493a5a9501630ff41ad2a261d6e1c93f73cfd1
Reviewed-on: https://chromium-review.googlesource.com/425529
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>