containers: support loopback devices
Containers that use a filesystem-in-a-file cannot be mounted directly.
They have to be added to a loop dev which is then mounted indirectly.

Add support for the standard "loop" keyword.  This will instruct the
container framework to set up the source via a loop device and then
mount that instead of the source file.

We also have to mount these outside of the userns namespace.  The
kernel does not allow mounting of any non-pseudo filesystems today
inside of a userns.  See the FS_USERNS_MOUNT logic in fs/namespace.c.

BUG=chromium:665622
TEST=started a container using a squashfs rootfs

Change-Id: I861f382c60d68267fdaae7c096178d8a21ca4c3d
Reviewed-on: https://chromium-review.googlesource.com/425899
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>