metrics: centralize client id look up
We provide limited sanity checking on the client id file, and we do so
in different places (differently).  Centralize the logic in the metrics
client so we don't duplicate the logic and we only have to robustify a
single code path.

I'm not a fan of exposing this ID more, but forcing people to duplicate
the logic is clearly worse in practice.

BUG=chromium:678365
TEST=precq passes

Change-Id: Ib8c16fa0b3dc7d5e42cfb3f81df29d99336613d6
Reviewed-on: https://chromium-review.googlesource.com/422851
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>