login: Use master configuration to select wallpaper files
Read the master configuration if available, and use this to select
model-specific wallpaper if available. Add a warning message if the files
are missing, to avoid confusion.

Signed-off-by: Simon Glass <sjg@chromium.org>
BUG=chromium:671391
TEST=Boot on reef, see that the model-specific wallpaper appears
Remove /usr/share/chromeos-config/config.dtb and see that the default
wallpaper appears (as would happen with USE=-uniboard), and an error is
logged:
2017-01-10T12:10:22.022037-07:00 ERR session_manager[3916]: [ERROR:cros_config.cc(95)] Could not read file /usr/share/chromeos-config/config.dtb

Change-Id: Ib5cf0669470d258436b7c9ca0b9cf6eccd782b84
Reviewed-on: https://chromium-review.googlesource.com/427139
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>