authpolicy: Prefer sAMAccountName for objectGUID search
Use sAMAccountName for searching for the objectGUID/account id and
fall back to userPrincipalName. This is what Windows and kinit seem
to do, i.e. if you kinit name@realm.com, it assumes that name is a
sAMAccountName first and only if there is no such sAMAccountName,
it assumes it's a UPN.

BUG=chromium:680454
TEST=Emerges, tested both cases with and without fallback.

Change-Id: Ic05f8b46e8f33b04c21a8d722ab6db27b007b31c
Reviewed-on: https://chromium-review.googlesource.com/427680
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>