authpolicy: Query workgroup from server
The previous code assumes that the workgroup is the first part of the
realm, i.e. realm = workgroup.domain.com. This is not true in general.
This CL queries the workgroup from the server instead. To handle the
rare case of workgroup renaming, the workgroup is queried once a
session and cached in the same way as the domain controller name.

BUG=chromium:680091
TEST=Emerges, tested with custom test code.

Change-Id: I070a6532384ce39d297455566b7e9dcd9959a72d
Reviewed-on: https://chromium-review.googlesource.com/427461
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Thiemo Nagel <tnagel@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>