authpolicy: Enable/fix seccomp filter failure logging
Previously, seccomp filter failures could not be logged since binaries
had to be executed without Minijail's preload library. This caused the
signal handler that logs failures to be wiped by execve since the handler
is installed before execve is called.

The reason the preload library could not be used is that authpolicyd uses
a saved uid (see setresuid) to switch to authpolicyd-exec in minijail.
The saved uid, however, is wiped by execve, so minijail can't switch
to it in the preload library.

This CL switches to authpolicyd-exec in authpolicy code before minijail
is run. This allows using the preload library and seccomp filter logging
works fine.

BUG=chromium:680082
TEST=Tested setting both flags, works as expected

Change-Id: If34e961fcbc37be111c066f6d32a84b256f5bba9
Reviewed-on: https://chromium-review.googlesource.com/427705
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>