login: set less restrictive value for perf_event_paranoid under dev-mode
/proc/sys/kernel/perf_event_paranoid controls the use of kernel perf
events system by users without the CAP_SYS_ADMIN capability. On systems
with Chrome OS kernel 3.14 or later, we now enable
CONFIG_SECURITY_PERF_EVENTS_RESTRICT in kernel, which sets the default
value of perf_event_paranoid to a more restrictive level "3", instead of
the previous default value "1". We already allow dev-mode only syscalls,
such as perf_event_open(), to be exposed to the Android container when
the system is in developer mode (CL:425454). This CL restores
perf_event_paranoid to the less restrictive level "1" when the system is
in developer mode in order to allow the following mapping:

     Chrome OS   |         Android        | Kernel
  ---------------+------------------------+-----------------------
   verified mode | security.perf_harden=1 | perf_event_paranoid=3
  ---------------+------------------------+-----------------------
  developer mode | security.perf_harden=0 | perf_event_paranoid=1

BUG=b:33450588
TEST=Tested the following:
- On samus-cheets (kernel 3.14):
  * Under developer mode, sign in a registered user, verify that
    /proc/sys/kernel/perf_event_paranoid is set to 1 by default.
  * Under verified mode, sign in a registered user, verify that
    /proc/sys/kernel/perf_event_paranoid is set to 3 by default.
- On a system with kernel 3.8, verify that
  /proc/sys/kernel/perf_event_paranoid is set to 1 by default in absence
  of the CONFIG_SECURITY_PERF_EVENTS_RESTRICT kernel option.

Change-Id: Ie78b276c3e4e0759f62171b6ed156b7594f28cd5
Reviewed-on: https://chromium-review.googlesource.com/428136
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>