power: power off/on Huawei ME936 modem on Kip during suspend/resume
CL:324106 modified powerd_suspend to power off the Huawei ME936 modem on
Blaze before the system goes to suspend and power the modem on after the
system resumes from suspend. The same modem is used on Kip and would be
affected by chrome-os-partner:48653. This CL ports CL:324106 to work on
Kip. Unlike Blaze that has a kernel accessible GPIO to toggle the modem
power, we need to toggle the modem power on Kip over EC using the
`ectool wireless` command, which introduces a dependency on
chromeos-base/ec-utils to chromeos-base/power_manager when building with
the cellular feature (i.e. USE=cellular).

BUG=chrome-os-partner:48653
CQ-DEPEND=CL:430071
TEST=Verified that the ME936 modem is powered off on Kip before the
     system goes into suspend and then powered on again after the system
     resumes from suspend.

Change-Id: I856e055e581f51a83b47cc00476c1883a9f26b58
Reviewed-on: https://chromium-review.googlesource.com/429931
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>