touch_keyboard: Add FakeKeyboard functionality
This CL adds the actual keyboard functionality to the touch_keyboard
module.  Previously it only supported FakeTouchpad's which is only
half of the functionality meant to be supplied.  Both the FakeKeyboard
and FakeTouchpad classes work in similar ways, leveraging evdev and
uinput to read events from a touch sensor and then convert them into
either a fake touchpad or fake keyboard.

Keyboard support is a work-in-progress, and still needs some changes
(several of which are marked with clear TODO's), but until this lands
the touch keyboard on Pbody won't work at all so it'd be good to get
this merged in now and follow up with tweaks and polish.  That way
devs and dogfooders will be able to type/etc without a USB external
keyboard.

BUG=chrome-os-partner:59138
TEST=I emerged and tested the touch_keyboard package on a Pbody.  The
keyboard worked on boot as well as the touchpad and I'm able to type
for the first time.

Change-Id: Id136d21d5ea7486e0118aacef140ae21611dfc66
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/430232
Reviewed-by: Jingkui Wang <jkwang@google.com>