container_utils: don't mount sysfs by default
The OCI spec confusingly says that certain filesystems must be
"made available" in containers, but it appears the onus is still
on the container's config.json to say whether or not it wants
those filesystems. However, to enforce a bare minimum of sanity,
we will sanitize the mount options for the sysfs and procfs mounts
if they are specified in the config.

BUG=chromium:698104
TEST=container starts when sysfs is a mount in config.json
CQ-DEPEND=CL:449057

Change-Id: I091534f26aef75419c788851722d30568ba7aa23
Reviewed-on: https://chromium-review.googlesource.com/431318
Commit-Ready: Stephen Barber <smbarber@chromium.org>
Tested-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>