crash: Join session keyring when collecting a user crash
Once ext4-based directory encryption is enabled, the encryption keys will be
stored in a session keyring created by init.  Since most processes get started
by init they will naturally inherit the keyring from init.  However,
crash_reporter is directly invoked by the kernel to record a core dump so it
doesn't automatically start with access to the keyring.

Have the crash_reporter attempt to join the session keyring before writing out
the core dump.  It may be the case that the session keyring does not exist,
which would imply that ext4 encryption is not enabled on that device so we don't
need to join the session keyring anyway.  In the case the core dump should still
be stored successfully as before.

BUG=chromium:672397
TEST=manual
On a cave with ext4 encryption enabled and on a peach_pit without ext4
encryption enabled, run the following steps:
- start python
- in a separate shell kill the python process with 'kill -11'
- grep crash_reporter /var/log/messages and confirm that the core dump was
  successfully written out
CQ-DEPEND=CL:433177,CL:433534

Change-Id: I76541950238d1a6f83e9a13ffd32254a468829fd
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/433098
Reviewed-by: Mike Frysinger <vapier@chromium.org>