libweave: fix test compilation
We copy in this gyp file to compile libweave, but there is a small
discrepancy between this one and the one in the libweave git repo
which leaves out one source file during FEATURES=test compilation.
This fixes the undefined vtable for weave::test::MockCommand.

In addition, the file containing the main function for the
test runner was not included in the gyp file, leading to another
undefined reference.

BUG=chromium:679477
TEST=FEATURES=test emerge-arkham libweave

Change-Id: Ic1349c550282efa39d844186e261d60ad85a490a
Reviewed-on: https://chromium-review.googlesource.com/433199
Commit-Ready: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>