authpolicy: Fetch GPO with sAMAccountName
Fixes policy fetch in case the user principal name (UPN) and sAMAccountName
don't match. Searching the objectGUID already handled this correctly, but
when the user logged in with the UPN, policy fetch would fail since it
requires the sAMAccountName. This CL properly retrieves the
sAMAccountName and uses it for policy fetch.

BUG=chromium:686231
TEST=Successfully fetched policy for an account with UPN != sAMAccountName

Change-Id: Ib83f9688e339e2f1d87cb4ea2bb0556b337f86ca
Reviewed-on: https://chromium-review.googlesource.com/435358
Commit-Ready: Lutz Justen <ljusten@chromium.org>
Tested-by: Lutz Justen <ljusten@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@chromium.org>