biod: update D-Bus API to send auth labels and get user enrollments
The D-Bus API is changed in two client facing ways:
1) GetEnrollments is now GetEnrollmentsForUser and requires that the
caller limits their query to a specific user ID
2) The Attempt signal replaces the list of user IDs with a map of user IDs
(keys) to lists of user assigned labels (values) for the purposes of
display of those labels back to the user.

The corresponding APIs interal to biod are updated along with the client
tool and fake biometric tool to facilitate testing.

TEST=biod_client_tool and fake_biometric_tool
BUG=chromium:629657

Change-Id: Id6818b90f4302b818f50040d3cdd4e3a6fe9f36f
Reviewed-on: https://chromium-review.googlesource.com/440430
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>