cryptohome: don't overwrite TpmState if dependencies don't change
For Attestation dependency a RemoverOwnerDependency is called on
every boot after successful PrepareForEnrollment to make sure
the dependency is removed even if the method failed originally.
Same is being introduced for the Nvram dependency.

The old logic was always storing the updated TpmState even if
the dependency flag was not set already and nothing changed.
The StoreTpmState operation is a bit heavy, with shredding and
rewriting the entire file, and may leave the file deleted if
interrupted at the wrong moment.
This change saves TpmState only when dependency flags changed.

BUG=chrome-os-partner:62689
TEST=unit tests

Change-Id: Ifcbcbc1aa621bc71d6991bcba38d82ad96b03783
Reviewed-on: https://chromium-review.googlesource.com/441191
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>