libchromeos-ui: chromium_command_builder: add flag for edge touch filtering
This CL adds the "--edge-touch-filtering" flag to the Chrome command
line if the "edge_touch_filtering" USE flag was set for this board. That
USE flag is found in /etc/ui_use_flags.txt, which is set based on the
USE flags for that board during build time.

CQ-DEPEND=CL:446131,CL:446034
BUG=chrome-os-partner:60856
TEST=I deployed libchromeos-ui to my Caroline.  I checked the command
line for chrome to see that the flag had been added when the USE flag
was added,but not when the USE flag was removed.

Change-Id: Id46ab64e43fc98709e63cbd2032139771c2ebaf1
Reviewed-on: https://chromium-review.googlesource.com/446616
Commit-Ready: Sean O'Brien <seobrien@chromium.org>
Tested-by: Sean O'Brien <seobrien@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>