libchromeos-ui: Re-enable native GpuMemoryBuffers.
https://crrev.com/6e8cd2e393e15f7e7f78a2ec0da4ee564da732a6 disabled native
GMB since we had many crash reports with mmap failing: crbug.com/629521.

Since we haven't been able to reproce the issue on a test device,
the plan is to re-enabled native GMBs on ToT with this CL after
crrev.com/2710183005 lands.
In this way we can get extra information with crash reports that could be
useful to debug the mmap issue.

BUG=chromium:629521
TEST=emerge-kevin libchromeos-ui

Change-Id: I1e728d12017f80d2ac13a89febe84f03e87e4d8d
Reviewed-on: https://chromium-review.googlesource.com/446161
Commit-Ready: Daniele Castagna <dcastagna@chromium.org>
Tested-by: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>