power: Modify dbus tool to control kb backlight
BUG=chromium:697988

The backlight_dbus_tool is used to control the screen
backlight intensity from the command line. As the
keyboard backlight uses a similar methodology and I
needed a command-line way to manipulate the keyboard
backlight, it seemed like a reasonable place to
add this functionality.

TEST=Built and ran command-line tool

Change-Id: Ib55983afb51eede4066a21c0e4de6a3084581a39
Reviewed-on: https://chromium-review.googlesource.com/448856
Commit-Ready: Nathan Baker <nathanb@lenovo-chrome.com>
Tested-by: Nathan Baker <nathanb@lenovo-chrome.com>
Reviewed-by: Dan Erat <derat@chromium.org>