shill: remove manager references to service being deregistered
When Manager deregisters a Service object, Manager may still hold a
reference to the Service object via
|Manager::last_default_physical_service_|, which defers the destruction
of the Service object until the scheduled SortServicesTask is executed
at a later point. By then, the Device object associated with the Service
object is already destroyed, which violates the assumption that the
Device object typically outlives the Service object.

As a Service object doesn't need to live beyond the time it gets
deregistered, this CL makes sure that Manager doesn't hold any more
reference to a Service object being deregistered.

BUG=chromium:699303
TEST=Run unit tests.
TEST=Manually unplug a USB Ethernet adapter and observe from the shill
log that the EthernetService object is destroyed before the
SortServicesTask is scheduled.

Change-Id: I74d9fd11ba178e28c9deb091d35c75af4a412ee9
Reviewed-on: https://chromium-review.googlesource.com/452791
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>