cryptohome: fix transient error logic for tpm2
Use the same IsTransient() logic to determine if the retry action
corresponds to a transient tpm error for TPM 2.0 case, as was used
for TPM 1.2.

IsTransient() is used to determine if a transient error occurred
when loading the cryptohome key, so that loading just needs to
be retried at a later time, instead of re-creating the key.

Map tpm device write/read errors to CommFailure retry action,
which is considered to be transient.

BUG=chromium:699240
TEST=Unit tests (new added)

Change-Id: I03ae89dec9a69a371edc2e0df700ec50f50c23fd
Reviewed-on: https://chromium-review.googlesource.com/452729
Commit-Ready: Andrey Pronin <apronin@chromium.org>
Tested-by: Andrey Pronin <apronin@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>