hal_adapter: close valid fds on camera_buffer_handle_t destruction
The call to native_handle_close() will blindly close all the |numFds|
fds in the native handle, no matter the fds are valid or not. This is
causing trouble because it may close fds such as 0, 1, or 2. Let's
make sure it only closes valid fds.

BUG=b:30824236
TEST=Stress test with arc-camera3-hal-usb and make sure preview works.

Change-Id: Ia9773496778e7771f3a798125cc7128784a2c05b
Reviewed-on: https://chromium-review.googlesource.com/453578
Commit-Ready: Ricky Liang <jcliang@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Heng-ruey Hsu <henryhsu@chromium.org>