midis: Add code for device monitoring.
This patch adds a DeviceManager class and sets up a thread
to watch for device connection/disconnection.

It also uses a combination of udev and ioctl (mostly udev), to extract
device information, and add it to a list of devices (maintained by the
device manager).

Devices are key-ed by a combination of their card number and device
number.

The same key is used to remove devices on disconnection.

BUG=chromium:701005
TEST=emerge-veyron_minnie midis
CQ-DEPEND=CL:456743

Change-Id: I0496f4e1f0ef070e40429bd724e84807334ccc79
Reviewed-on: https://chromium-review.googlesource.com/456764
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>