shill: populate a missing error in Manager::ConfigureServiceForProfile
When Manager::ConfigureServiceForProfile() creates and configures a
temporary service, it decides not to return the unusable service but
forgets to populate an error, which will eventually trigger a assertion
in ChromeosManagerDBusAdaptor::ConfigureServiceForProfile(). This CL
ensures that Manager::ConfigureServiceForProfile populates an error
under such condition.

BUG=chromium:701315
TEST=Run unit tests.

Change-Id: Id990de604982cffcc68dff2899ac9859935625bb
Reviewed-on: https://chromium-review.googlesource.com/455256
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>